+++ backend/src/auth/kakao/auth.kakao.controller.ts
@@ -2,10 +2,10 @@ import { Controller, Get, Logger, Res, UseGuards } from '@nestjs/common';
 import { Response } from 'express';
 import { User } from 'src/decorator/user.decorator';
 import { UserSessionDto } from 'src/dto/user.session.dto';
-import { AuthService } from './auth.service';
-import { JwtAuthGuard } from './jwt/guard/jwt.auth.guard';
-import { JWTSignGuard } from './jwt/guard/jwt.sign.guard';
-import { KakaoGuard } from './kakao/guard/kakao.guard';
+import { AuthService } from '../auth.service';
+import { JwtAuthGuard } from '../jwt/guard/jwt.auth.guard';
+import { JWTSignGuard } from '../jwt/guard/jwt.sign.guard';
+import { KakaoGuard } from './guard/kakao.guard';
 
 @Controller('auth/kakao')
 export class AuthKakaoController {
@@ -24,13 +24,6 @@ export class AuthKakaoController {
   @UseGuards(KakaoGuard, JWTSignGuard)
   async loginCallback(@Res() res: Response) {
     this.logger.log('login success!');
-    return res.redirect('/');
-  }
-
-  @Get('/logout')
-  @UseGuards(JwtAuthGuard)
-  async logout(@Res() res: Response, @User() user: UserSessionDto) {
-    await this.authService.logoutKakao(res, user);
     return res.redirect('/');
   }
 }
