+++ backend/src/auth/naver/auth.naver.controller.ts
@@ -2,10 +2,10 @@ import { Controller, Get, Logger, Res, UseGuards } from "@nestjs/common";
 import { Response } from "express";
 import { User } from "src/decorator/user.decorator";
 import { UserSessionDto } from "src/dto/user.session.dto";
-import { AuthService } from "./auth.service";
-import { JwtAuthGuard } from "./jwt/guard/jwt.auth.guard";
-import { JWTSignGuard } from "./jwt/guard/jwt.sign.guard";
-import { NaverGuard } from "./naver/guard/naver.guard";
+import { AuthService } from "../auth.service";
+import { JwtAuthGuard } from "../jwt/guard/jwt.auth.guard";
+import { JWTSignGuard } from "../jwt/guard/jwt.sign.guard";
+import { NaverGuard } from "./guard/naver.guard";
 
 
 @Controller('auth/naver')
@@ -25,13 +25,6 @@ export class AuthNaverController {
   @UseGuards(NaverGuard, JWTSignGuard)
   async loginCallback(@Res() res: Response) {
     this.logger.log('login success!');
-    return res.redirect('/');
-  }
-
-  @Get('/logout')
-  @UseGuards(JwtAuthGuard)
-  async logout(@Res() res: Response, @User() user: UserSessionDto) {
-    await this.authService.logout(res, user);
     return res.redirect('/');
   }
 }
